+++ sdk/src/store.rs
@@ -1669,14 +1669,11 @@ impl Store {
                                 "ingredient hash does not match found ingredient".to_string(),
                             ),
                         )?;
-                        return Err(Error::HashMismatch(
-                            "ingredient hash does not match found ingredient".to_string(),
-                        )); // hard stop regardless of StatusTracker mode
                     }
 
-                    // if manifest hash did not match and this is a V2 or greater claim then we
+                    // if manifest hash did not match because of redaction and this is a V2 or greater claim then we
                     // must try the signature validation method before proceeding
-                    if !manifests_match && ingredient_version > 1 {
+                    if !manifests_match && has_redactions && ingredient_version > 1 {
                         let claim_signature =
                             ingredient_assertion.signature().ok_or_else(|| {
                                 log_item!(
@@ -1724,10 +1721,6 @@ impl Store {
                                     "ingredient claimSignature mismatch".to_string(),
                                 ),
                             )?;
-                            return Err(Error::HashMismatch(
-                                "ingredient signature box hash does not match found ingredient"
-                                    .to_string(),
-                            )); // hard stop regardless of StatusTracker mode
                         }
                     }
 
@@ -1737,19 +1730,16 @@ impl Store {
                         Claim::verify_hash_binding(ingredient, asset_data, svi, validation_log)?;
                     }
 
-                    // if manifest hash did not match we continue on to do a full claim validation
-                    if !manifests_match {
-                        Claim::verify_claim(
-                            ingredient,
-                            asset_data,
-                            svi,
-                            check_ingredient_trust,
-                            &store.ctp,
-                            validation_log,
-                            http_resolver,
-                            settings,
-                        )?;
-                    }
+                    Claim::verify_claim(
+                        ingredient,
+                        asset_data,
+                        svi,
+                        check_ingredient_trust,
+                        &store.ctp,
+                        validation_log,
+                        http_resolver,
+                        settings,
+                    )?;
 
                     // recurse nested ingredients
                     Store::ingredient_checks(
@@ -1856,7 +1846,7 @@ impl Store {
 
                     // allow the extra ingredient trust checks
                     // these checks are to prevent the trust spoofing
-                    let check_ingredient_trust = settings.verify.check_ingredient_trust;
+                    let check_ingredient_trust = settings.verify.verify_trust;
 
                     // get the 1.1-1.2 box hash
                     let ingredient_hashes = store.get_manifest_box_hashes(ingredient);
@@ -1903,14 +1893,11 @@ impl Store {
                                 "ingredient hash does not match found ingredient".to_string(),
                             ),
                         )?;
-                        return Err(Error::HashMismatch(
-                            "ingredient hash does not match found ingredient".to_string(),
-                        )); // hard stop regardless of StatusTracker mode
                     }
 
                     // if manifest hash did not match and this is a V2 or greater claim then we
                     // must try the signature validation method before proceeding
-                    if !manifests_match && ingredient_version > 1 {
+                    if !manifests_match && has_redactions && ingredient_version > 1 {
                         let claim_signature =
                             ingredient_assertion.signature().ok_or_else(|| {
                                 log_item!(
@@ -1958,10 +1945,6 @@ impl Store {
                                     "ingredient claimSignature mismatch".to_string(),
                                 ),
                             )?;
-                            return Err(Error::HashMismatch(
-                                "ingredient signature box hash does not match found ingredient"
-                                    .to_string(),
-                            )); // hard stop regardless of StatusTracker mode
                         }
                     }
 
@@ -1971,20 +1954,17 @@ impl Store {
                         Claim::verify_hash_binding(ingredient, asset_data, svi, validation_log)?;
                     }
 
-                    // if manifest hash did not match we continue on to do a full claim validation
-                    if !manifests_match {
-                        Claim::verify_claim_async(
-                            ingredient,
-                            asset_data,
-                            svi,
-                            check_ingredient_trust,
-                            &store.ctp,
-                            validation_log,
-                            http_resolver,
-                            settings,
-                        )
-                        .await?;
-                    }
+                    Claim::verify_claim_async(
+                        ingredient,
+                        asset_data,
+                        svi,
+                        check_ingredient_trust,
+                        &store.ctp,
+                        validation_log,
+                        http_resolver,
+                        settings,
+                    )
+                    .await?;
 
                     // recurse nested ingredients
                     Box::pin(Store::ingredient_checks_async(